Aggiornamento dei paramentri: da liste di Float a liste di Double

+++ JFML_EntryPoint/src/jfml_entrypoint/ExtendedAnYaRuleType.java
@@ -27,7 +27,7 @@ public class ExtendedAnYaRuleType extends AnYaRuleType{
 		super(name,weight);
 	}
 	
-	public float andFunction(ArrayList<Float> degreesList){
+	public float andFunction(ArrayList<Double> degreesList){
 		float[] degrees = new float[degreesList.size()];
 		for (int i = 0; i < degreesList.size(); i++) {
 			degrees[i] = degreesList.get(i).floatValue();
@@ -35,7 +35,7 @@ public class ExtendedAnYaRuleType extends AnYaRuleType{
 		return this.and(degrees);
 	}
 	
-	public float andFunction(String andMethod, ArrayList<Float> degreesList){
+	public float andFunction(String andMethod, ArrayList<Double> degreesList){
 		float[] degrees = new float[degreesList.size()];
 		for (int i = 0; i < degreesList.size(); i++) {
 			degrees[i] = degreesList.get(i).floatValue();
@@ -43,7 +43,7 @@ public class ExtendedAnYaRuleType extends AnYaRuleType{
 		return this.and(andMethod, degrees);
 	}
 	
-	public float orFunction(ArrayList<Float> degreesList){
+	public float orFunction(ArrayList<Double> degreesList){
 		float[] degrees = new float[degreesList.size()];
 		for (int i = 0; i < degreesList.size(); i++) {
 			degrees[i] = degreesList.get(i).floatValue();
@@ -51,7 +51,7 @@ public class ExtendedAnYaRuleType extends AnYaRuleType{
 		return this.or(degrees);
 	}
 	
-	public float orFunction(String andMethod, ArrayList<Float> degreesList){
+	public float orFunction(String andMethod, ArrayList<Double> degreesList){
 		float[] degrees = new float[degreesList.size()];
 		for (int i = 0; i < degreesList.size(); i++) {
 			degrees[i] = degreesList.get(i).floatValue();
